Modificación de mapeos
Desde la parte de servidor, he modificado las variables de los mapeos y la creación de funciones con el fin de haber conseguido eliminar y crear posts, by krimo

+++ practicaFinal/src/main/java/com/practicafinal/team/manager/PostManager.java
@@ -23,7 +23,7 @@ public class PostManager {
         return postRepository.findById(theId);
     }
 
-    public void deletePlaneta(int theId) {
+    public void deletePlaneta(long theId) {
         postRepository.deleteById(theId);
     }
 }
